Move check support image to `files` pkg
