+++ internal/files/images.go
@@ -1,4 +1,4 @@
-package fsprov
+package files
 
 import (
 	"path/filepath"
@@ -6,7 +6,7 @@ import (
 )
 
 // Check if image file extension is supported.
-func isSupportedImg(filename string) bool {
+func IsSupportedImg(filename string) bool {
 	ext := filepath.Ext(filename)
 	ext = strings.ToLower(ext)
 	return ext == ".jpg" || ext == ".png" || ext == ".jpeg" || ext == ".webp"
